collateral node enforce 2022
Changed masternode enforcement height to block 2022. Originally was 4500 for D with a 30 sec block time, was changed to 20000 and then 80000 on account of FS not starting up. I plan to get FS (collateral node/masternode) working before 2022 -punny-. Set to 2022 for the current year and is almost the same amount of real time as a 4500 height with 30 block time. Note Rupee Evolution block time should be 60 seconds. First rebrand of fortunastakes to collateral node in //notes.